requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
sqlmodel==0.0.8
pandas==2.2.3
celery==5.2.10
redis==4.6.0
python-dotenv==1.0.0
